flask-server/requirements.txt
backoff
flask
Flask-SQLAlchemy
mysql
requests
sqlalchemy
opentelemetry-instrumentation
opentelemetry-sdk
opentelemetry-ext-jaeger